webrtc: Add support for add and remove track
WebRTC has support to add and remove tracks while a connection is
established. This change adds the API to trigger that behavior when a
client application wants to implement this. Furthermore changing the
local stream now automatically triggers a remove and add of all the
tracks in the local stream.