kwm: Properly keep Promise chain on reconnect
This change ensures that the initial connect Promise is properly
resolved or thrown when the first connect failed and internal retry is
enabled. The outer promise now is passed through properly and is
resolved when the reconnecting code managed to establish the
connection. Before this promise was never resolved as soon as the
connect code entered the reconnector.