Add webrtc payload version
To avoid trying to connect with incompatible clients this change adds a
version marker to WebRTC payload data. This marker is used to ignore
messages from older clients which might be not compatible. The version
is a simple date stamp which should be increased whenever an
incompatible change is made within the webrtc payload data or behavior.