Add doReject WebRTC function
To distinguish between hangup (which is when a call is already accepted)
and reject (rejecting to accept a call), a new api function doReject is
added to the WebRTC manager. Call this function instead of hangup to
reject pending calls to ensure that other ringing sessions of the same
user are properly notified about the call rejection and also stop
ringing.