Restore lost WeakMap before removing a track
For unknown reason, SimplePeer looses its stored references to
RTCRtpSender once in a while. When this happens and a track for that
sender is removed (e.g. on mute), the connection gets destroyed since
that counts as fatal exception within SimplePeer. This change checks if
the sender is still inside the internal SimplePeer data structure and
tries to revoder it from the PeerConnection senders.