Add pcid to webrtc signal messages
To identifiy unique peer connections per client, this change adds the
`pcid` field to RTMTypeWebRTC messages to hold a unique value for each
client peer connection. The remote pcid is stored at the peer record and
compared. Whenever it changes, the local peer connection is destroyed as
well, so that both ends start clean in case one peer had to recreate its
connection.