Use Plan B SDP by default (Chrome)
Chrome is transitioning to the WebRTC unified plan. While this plan is
standard conformant and we should use it in the future, it also breaks
backwards compatibility with older Chrome (<72) which only handle Plan B
and if multiple streams of the same type are used. Thus for a transition
phase on our end, we default to Plan B to retain support for older
Chromes.

Reference: https://webrtc.org/web-apis/chrome/unified-plan/