Avoid loosing required negotiation when not the initator
I noticed that sometimes a newly added track is not transmitted to the
remote peer. This happend when the track is added while the connection
is not yet stable and when the local peer is not the initiator.

This change updates to forked simple-peer to include the fix without
having to wait for a upstream release.

Reference: https://github.com/feross/simple-peer/pull/583