Avoid channel updates after hangup
On hangup, other peers might also hang up and thus channel updates are
triggered by the server. If a channel update is received out of bounds
(not from a reply of our own) and the channel in the update is different
from the local one such messages are either for a different channel or
from the one which was just removed locally. This cures a race
condition, fixing the issue that a local channel was still present after
hangup.